lint-workflow-v2/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bitwarden_workflow_linter"
dynamic = ["version"]
authors = ["Bitwarden Inc"]
description = "Custom GitHub Action Workflow Linter"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]

dependencies = [
	"annotated-types==0.6.0",
	"dataclasses-json==0.6.4",
	"marshmallow==3.20.2",
	"mypy-extensions==1.0.0",
	"packaging==23.2",
	"pydantic==2.6.0",
	"pydantic-core==2.16.1",
	"pyyaml==6.0.1",
	"ruamel.yaml==0.18.5",
	"ruamel.yaml.clib==0.2.8",
	"typing-extensions==4.9.0",
	"typing-inspect==0.9.0",
	"urllib3==2.2.0",
]

[project.urls]
Homepage = "https://github.com/bitwarden/gh-actions/tree/main/lint-workflow-v2"
Issues = "https://github.com/bitwarden/gh-actions/issues"

[project.scripts]
bwwl = "bitwarden_workflow_linter.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/bitwarden_workflow_linter"]

[tool.hatch.version]
path = "src/bitwarden_workflow_linter/__about__.py"
